allow to use different paths to serve normal get/set/del requets and the internal __stats__, __index__ and __health__ pages
-B can be used on the cmdline to define an alternative administrative path
and same can be done using the 'baseadminpath' directive in the config file.

If the administrative path and the base one match, the behaviour will be exactly the same
it was before (with the special keys being reserved and not usable for caching purposes but only
to retreive the administrative information).
If they differ, any key can be used for caching purposes while administrative pages
will be accessible via the different path